+++ eslint.config.mjs
@@ -1,0 +1,49 @@
+import typescriptEslint from "@typescript-eslint/eslint-plugin";
+import tsParser from "@typescript-eslint/parser";
+import path from "node:path";
+import { fileURLToPath } from "node:url";
+import js from "@eslint/js";
+import { FlatCompat } from "@eslint/eslintrc";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+const compat = new FlatCompat({
+  baseDirectory: __dirname,
+  recommendedConfig: js.configs.recommended,
+  allConfig: js.configs.all
+});
+
+export default [...compat.extends(
+  "plugin:@typescript-eslint/recommended",
+  "plugin:@typescript-eslint/recommended-requiring-type-checking",
+  "plugin:jest/recommended",
+  "prettier",
+), {
+  plugins: {
+    "@typescript-eslint": typescriptEslint,
+  },
+
+  languageOptions: {
+    parser: tsParser,
+    ecmaVersion: 2018,
+    sourceType: "module",
+
+    parserOptions: {
+      project: ["./tsconfig.json", "./tsconfig.spec.json"],
+    },
+  },
+
+  rules: {
+    "@typescript-eslint/no-explicit-any": "off",
+    "@typescript-eslint/no-unused-vars": ["error", {
+      args: "none",
+    }],
+    "@typescript-eslint/no-unsafe-return": "off",
+    "@typescript-eslint/no-unsafe-assignment": "off",
+    "@typescript-eslint/no-unsafe-call": "off",
+    "@typescript-eslint/no-unsafe-member-access": "off",
+    "@typescript-eslint/no-unsafe-argument": "off",
+    "@typescript-eslint/no-require-imports": "off",
+    "@typescript-eslint/no-redundant-type-constituents": "off",
+  },
+}];