+++ src/RoutingControllers.ts
@@ -38,7 +38,10 @@ export class RoutingControllers<T extends BaseDriver> {
   // Constructor
   // -------------------------------------------------------------------------
 
-  constructor(private driver: T, private options: RoutingControllersOptions) {
+  constructor(
+    private driver: T,
+    private options: RoutingControllersOptions,
+  ) {
     this.parameterHandler = new ActionParameterHandler(driver);
     this.metadataBuilder = new MetadataBuilder(options);
   }
@@ -50,8 +53,8 @@ export class RoutingControllers<T extends BaseDriver> {
   /**
    * Initializes the things driver needs before routes and middleware registration.
    */
-  initialize(): this {
-    this.driver.initialize();
+  initialize(options: RoutingControllersOptions): this {
+    this.driver.initialize(options);
     return this;
   }