+++ src/index.ts
@@ -155,19 +155,19 @@ export function createServer<T extends BaseDriver>(driver: T, options?: RoutingC
  */
 export function createExecutor<T extends BaseDriver>(driver: T, options: RoutingControllersOptions = {}): void {
   // import all controllers and middlewares and error handlers (new way)
-  let controllerClasses: Newable[];
+  let controllerClasses: Newable[] | undefined;
   if (options && options.controllers && options.controllers.length) {
     controllerClasses = (options.controllers as any[]).filter(controller => controller instanceof Function);
     const controllerDirs = (options.controllers as any[]).filter(controller => typeof controller === 'string');
     controllerClasses.push(...importClassesFromDirectories(controllerDirs));
   }
-  let middlewareClasses: Newable[];
+  let middlewareClasses: Newable[] | undefined;
   if (options && options.middlewares && options.middlewares.length) {
     middlewareClasses = (options.middlewares as any[]).filter(controller => controller instanceof Function);
     const middlewareDirs = (options.middlewares as any[]).filter(controller => typeof controller === 'string');
     middlewareClasses.push(...importClassesFromDirectories(middlewareDirs));
   }
-  let interceptorClasses: Newable[];
+  let interceptorClasses: Newable[] | undefined;
   if (options && options.interceptors && options.interceptors.length) {
     interceptorClasses = (options.interceptors as any[]).filter(controller => controller instanceof Function);
     const interceptorDirs = (options.interceptors as any[]).filter(controller => typeof controller === 'string');
@@ -214,7 +214,7 @@ export function createExecutor<T extends BaseDriver>(driver: T, options: Routing
 
   // next create a controller executor
   new RoutingControllers(driver, options)
-    .initialize()
+    .initialize(options)
     .registerInterceptors(interceptorClasses)
     .registerMiddlewares('before', middlewareClasses)
     .registerControllers(controllerClasses)