+++ test/functional/json-controller-methods.spec.ts
@@ -77,14 +77,14 @@ describe(``, () => {
             name: 'Umed',
           };
         }
-        @Get(/\/categories\/[\d+]/)
+        @Get('/categories/:id')
         getCategoryById() {
           return {
             id: 1,
             name: 'People',
           };
         }
-        @Get('/posts/:id(\\d+)')
+        @Get('/posts/:id')
         getPostById() {
           return {
             id: 1,
@@ -218,41 +218,10 @@ describe(``, () => {
       });
     });
 
-    it('route should work with regexp parameter', async () => {
-      expect.assertions(3);
-      const response = await axios.get('/categories/1');
-      expect(response.status).toEqual(HttpStatusCodes.OK);
-      expect(response.headers).toHaveProperty('content-type', 'application/json; charset=utf-8');
-      expect(response.data).toEqual({
-        id: 1,
-        name: 'People',
-      });
-    });
-
-    it('should respond with 404 when regexp does not match', async () => {
+    it('should respond with 404 when route does not match', async () => {
       expect.assertions(1);
       try {
-        await axios.get('/categories/b1');
-      } catch (err: any) {
-        expect(err.response.status).toEqual(HttpStatusCodes.NOT_FOUND);
-      }
-    });
-
-    it('route should work with string regexp parameter', async () => {
-      expect.assertions(3);
-      const response = await axios.get('/posts/1');
-      expect(response.status).toEqual(HttpStatusCodes.OK);
-      expect(response.headers).toHaveProperty('content-type', 'application/json; charset=utf-8');
-      expect(response.data).toEqual({
-        id: 1,
-        title: 'About People',
-      });
-    });
-
-    it('should respond with 404 when regexp does not match', async () => {
-      expect.assertions(1);
-      try {
-        await axios.get('/posts/U');
+        await axios.get('/unknown/url');
       } catch (err: any) {
         expect(err.response.status).toEqual(HttpStatusCodes.NOT_FOUND);
       }